writing SimpleScore

+++ src/main/java/uk/ac/bris/cs/scotlandyard/ui/ai/Score.java
@@ -3,5 +3,5 @@ package uk.ac.bris.cs.scotlandyard.ui.ai;
 import uk.ac.bris.cs.scotlandyard.model.Board;
 
 public interface Score {
-    public Integer score(Board.GameState gameState);
+    public float score(Board.GameState gameState);
 }
